aws/AWS-CDK/DMS/amazon-dms-postgres-aurora-to-rds-aurora-cdc/requirements.txt
aws-cdk-lib==2.69.0
boto3==1.24.78

# Local development tools
flake8
python-dotenv==0.20.0
black